--- a/secio/src/crypto/ring_impl.rs
+++ b/secio/src/crypto/ring_impl.rs
@@ -96,29 +96,23 @@ impl RingAeadCipher {
             .len()
             .checked_sub(self.cipher_type.tag_size())
             .ok_or(SecioError::FrameTooShort)?;
-        let mut output = Vec::with_capacity(output_len);
-        let mut buf = Vec::with_capacity(
-            self.cipher_type
-                .tag_size()
-                .checked_add(input.len())
-                .ok_or(SecioError::InvalidMessage)?,
-        );
+        let mut buf = Vec::with_capacity(input.len());
 
         unsafe {
-            output.set_len(output_len);
             buf.set_len(input.len());
         }
         buf.copy_from_slice(input);
 
         if let RingAeadCryptoVariant::Open(ref mut key) = self.cipher {
             match key.open_in_place(Aad::empty(), &mut buf) {
-                Ok(out_buf) => output.copy_from_slice(out_buf),
+                Ok(_) => (),
                 Err(e) => return Err(e.into()),
             }
         } else {
             unreachable!("encrypt is called on a non-open cipher")
         }
-        Ok(output)
+        buf.truncate(output_len);
+        Ok(buf)
     }
 }