in place decrypt on p2p bench

before：
```
After counting 10 cycles, estimated total time spent on task "10mb_benchmark_with_secio" is 2.6936522s                                 
task name: 10mb_benchmark_with_secio                                                                                                   
cycles: 100                                                                                                                            
total cost: 3.628672348s                                                                                                               
average: 36.286723ms                                                                                                                   
median: 45.029141ms                                                                                                                    
max: 52.898023ms                                                                                                                       
min: 20.852433ms
```

after：
```
After counting 10 cycles, estimated total time spent on task "10mb_benchmark_with_secio" is 2.1110216s
task name: 10mb_benchmark_with_secio
cycles: 100
total cost: 2.433076186s
average: 24.330761ms
median: 24.386613ms
max: 55.059342ms
min: 19.604614ms
```